backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
pydantic-settings==2.7.1
python-jose[cryptography]==3.3.0
bcrypt==4.1.3
alembic==1.14.1
python-multipart==0.0.20
eval-type-backport==0.3.1
httpx==0.28.1
python-dateutil==2.8.2
aiofiles==23.2.1
